+++ tests/tinytest/test_count_lazy.R
@@ -79,4 +79,36 @@ expect_equal_lazy(
   TRUE
 )
 
+# message if overwriting variable
+
+test2 <- test |>
+  mutate(n = 1)
+
+test3 <- test2 |>
+  mutate(nn = 1)
+
+expect_message(
+  test2 |> count(n),
+  "Storing counts in `nn`, as `n` already present in input."
+)
+
+expect_message(
+  test2 |> add_count(cyl),
+  "Storing counts in `nn`, as `n` already present in input."
+)
+
+expect_message(
+  test3 |> add_count(cyl),
+  "Storing counts in `nnn`, as `n` already present in input."
+)
+
+expect_equal_lazy(
+  test2 |>
+    add_count(cyl, name = "n") |>
+    pull(n),
+  test |>
+    add_count(cyl, name = "n") |>
+    pull(n)
+)
+
 Sys.setenv('TIDYPOLARS_TEST' = FALSE)